Move check-flow utils methods to loadflow-validation module
