+++ iidm-api/src/test/java/eu/itesla_project/iidm/network/CoverageTest.java
@@ -1,0 +1,16 @@
+/**
+ * Copyright (c) 2017, RTE (http://www.rte-france.com)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package eu.itesla_project.iidm.network;
+
+import org.junit.Test;
+
+public class CoverageTest {
+
+    @Test
+    public void testNothing() {
+    }
+}
