+++ loadflow-validation/src/main/java/eu/itesla_project/loadflow/validation/CheckFlowsTool.java
@@ -24,7 +24,6 @@ import eu.itesla_project.computation.local.LocalComputationManager;
 import eu.itesla_project.iidm.import_.Importers;
 import eu.itesla_project.iidm.network.Network;
 import eu.itesla_project.iidm.network.StateManager;
-import eu.itesla_project.iidm.network.util.Networks;
 import eu.itesla_project.loadflow.api.LoadFlow;
 import eu.itesla_project.loadflow.api.LoadFlowFactory;
 import eu.itesla_project.loadflow.api.LoadFlowParameters;
